Fix review comments.
